[Features] Small refactor, weather and URL
The actions are now defined in the constants file. An action consists
of a label, a hotkey, and an actionFile. The callback handler was
created to invoke the aproperiate handler inside the actionFile passing
the correct parameters. This should remain in the controler so we can
have custom parameters providade by the UI in the future.